changed the study_id col to REAL in the individual table to match the change I made in a previous commit that changed study_id col to REAL in study table, changed syntax in clean_study.r to replace depricated syntax

+++ db/create_db.sql
@@ -60,7 +60,7 @@ CREATE TABLE IF NOT EXISTS `individual` (
 	`individual_id`	INTEGER PRIMARY KEY,
 	`local_identifier`	TEXT,
 	`nick_name`	TEXT,
-	`study_id`	INTEGER,
+	`study_id`	REAL,
 	`ring_id`	TEXT,
 	`sex`	TEXT,
   `taxon_id`	INTEGER,
